+++ lib/calendar_assistant/calendar_assistant.rb
@@ -38,7 +38,6 @@ class CalendarAssistant
     @event_repository_factory = event_repository_factory
     @event_repositories = {} # calendar_id → event_repository
     @event_predicates = PredicateCollection.build(config.must_be, config.must_not_be)
-    @emoji_worldmap = Array(@config.setting(CalendarAssistant::Config::Keys::Settings::LOCATION_ICONS)).first
   end
 
   def in_env &block
@@ -96,7 +95,11 @@ class CalendarAssistant
     deleted_events = []
     modified_events = []
 
-    event = event_repository.create(transparency: CalendarAssistant::Event::Transparency::TRANSPARENT, start: range.first, end: range.last , summary: "#{@emoji_worldmap} #{location}")
+    event = event_repository.create(
+      transparency: CalendarAssistant::Event::Transparency::TRANSPARENT,
+      start: range.first, end: range.last,
+      summary: "#{Event.location_event_prefix(@config)}#{location}"
+    )
 
     existing_event_set.events.each do |existing_event|
       if existing_event.start_date >= event.start_date && existing_event.end_date <= event.end_date