+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog for Calendar Assistant
 
+## unreleased
+
+Features:
+
+- Allow location event visibility to be set. [#17]
+- Allow command-specific settings. [#17]
+- Support a `nickname` setting and use it in `location-set` to uniquely identify you in an event summmary. [#85]
+
+Breaking Changes:
+
+- Remove support for multiple location icons
+
 ## v0.7.0 / 2019-01-15
 
 Features: